requirements-fast.txt
face-recognition==1.3.0
dlib==19.24.2
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
numpy==1.24.3
opencv-python==4.9.0.80
pillow==10.2.0
